라인업 이미지는 좌우로 슬라이드 하는 형태로 움직일 수 있다
